github-package/backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pydantic>=2.6.4

# Database
SQLAlchemy>=2.0.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0
python-dateutil>=2.8.0
chardet>=5.2.0

# File handling
python-multipart>=0.0.9

# HTTP client
requests>=2.31.0

# Development and testing
pytest>=8.0.0

# Optional: for production deployment
# gunicorn>=21.0.0
# psycopg2-binary>=2.9.0  # For PostgreSQL